requirements.txt
streamlit
requests
python-dotenv
Pillow
plotly
openai-whisper
transformers
torch
scipy
google-generativeai
audio-recorder-streamlit
